move footer logic and add gallery media queries
