+++ src/App.css
@@ -3,15 +3,8 @@
     background-image: linear-gradient(0deg, #35b5ff 0%, #9ffff4 100%);
 }
 
-.footer {
-    height:200px;
-    background-color: #0093E9;
-    background-image: linear-gradient(0deg, #35b5ff 100%, #9ffff4 0%);
-}
-
 .container {
     display: flex;
     align-items: center;
     justify-content: center;
-    width: 90%;
   }
