+++ src/components/Footer/Footer.css
@@ -1,0 +1,5 @@
+.footer {
+    height:200px;
+    background-color: #0093E9;
+    background-image: linear-gradient(0deg, #35b5ff 100%, #9ffff4 0%);
+}
