+++ src/components/Gallery/Gallery.css
@@ -7,3 +7,18 @@
   justify-content: center;
   padding: 2rem;
   }
+
+
+  @media (max-width: 800px) {
+    .gallery {
+      grid-template-columns: repeat(2, 1fr);
+    }
+  }
+
+
+  @media (max-width: 600px) {
+    .gallery {
+      grid-template-columns: repeat(1, 1fr);
+    }
+  }
+
