+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as Gallery } from "./Gallery/Gallery";
 export { default as Navbar } from "./Navigation/Navigation";
+export { default as Footer } from "./Footer/Footer";
